Finish first draft of parser
The parser can now read in properly formatted text and parse it into a
valid board. Unfortunately, I just realized that it panics if the
text has an issue. This is because the parser does not assert that the
board text is that last thing in the file, so if it encounters
something it doesn't expect it simply returns early and does not
"consume" the rest of the text. Then, when we try to parse it into an
`Array2D`, there is a dimension mismatch.

In good news, the solver can solve 8x8 boards practically instantly on
my computer.